+++ xrpl/common/constants.go
@@ -13,4 +13,7 @@ const (
 	DefaultTimeout            = 10 * time.Second
 	DefaultFeeCushion float32 = 1.2
 	DefaultMaxFeeXRP  float32 = 2
+
+	// 5 seconds default timeout
+	DefaultTimeout = 5 * time.Second
 )