+++ xrpl/websocket/queries_test.go
@@ -1431,7 +1431,7 @@ func TestClient_GetLedger(t *testing.T) {
 			}},
 			expected: &ledgerqueries.Response{
 				Ledger: ledgertypes.BaseLedger{
-					LedgerIndex:         common.LedgerIndex(14380380),
+					LedgerIndex:         14380380,
 					TotalCoins:          types.XRPCurrencyAmount(99999999999999997),
 					ParentHash:          "ABC123",
 					TransactionHash:     "DEF456",