+++ CHANGELOG.md
@@ -7,20 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.1.6]
 
+### Added
+
+#### xrpl
+
+- Configurable timeout for the RPC client. New default timeout of 5 seconds instead of 1 second.
+
 ### Fixed
 
 #### xrpl
 
 - Updates some fields in AccountSet and Payment related transactions to a pointer to allow 0 or "" values. For example:
+
   - `DestinationTag`
   - `TickSize`
   - `Domain`
   - `WalletLocator`
   - `WalletSize`
   - `TransferRate`
-
 - Adds more tests for setting some `asf` flags in `AccountSet`.
 - Fixed `Transaction` field in `account_tx` response.
+- Fixed `Ledger` field in `ledger` response. LedgerIndex is now an uint32 instead of a string.
 
 ## [v0.1.5]